src/requirements.txt
pymysql>=1.1.0
pyyaml>=6.0